# cmd/management/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Source-of-truth API for managing GO Feature Flag flagsets, flags, teams, and audit history.",
        "title": "GO Feature Flag Management API",
        "contact": {},
        "version": "0.1.0"
    },
    "basePath": "/api/v1",
    "paths": {
...
        "/flagsets/{id}": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "flagsets"
                ],
                "summary": "Get a flagset",
                "parameters": [
                    {
                        "type": "string",
                        "description": "flagset id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/model.APIResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/model.Flagset"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/model.APIResponse"
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "flagsets"
                ],
                "summary": "Delete a flagset",
                "parameters": [
                    {
                        "type": "string",
                        "description": "flagset id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.APIResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/model.APIResponse"
                        }
                    }
                }
            },
            "patch": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "flagsets"
                ],
                "summary": "Update a flagset",
                "parameters": [
                    {
                        "type": "string",
                        "description": "flagset id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "fields to update",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.UpdateFlagsetRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/model.APIResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/model.Flagset"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/model.APIResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/model.APIResponse"
                        }
                    }
                }
            }
        },
        "/flagsets/{id}/api-keys": {
            "post": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "flagsets"
                ],
                "summary": "Create a new API key for a flagset (raw key returned ONCE)",
                "parameters": [
                    {
                        "type": "string",
                        "description": "flagset id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/model.APIResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/model.CreateAPIKeyResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/model.APIResponse"
                        }
                    }
                }
            }
        },
        "/flagsets/{id}/api-keys/{keyHash}": {
            "delete": {
                "tags": [
                    "flagsets"
                ],
                "summary": "Revoke a flagset API key by hash",
                "parameters": [
                    {
                        "type": "string",
                        "description": "flagset id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "key hash",
                        "name": "keyHash",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.APIResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/model.APIResponse"
                        }
                    }
                }
            }
        },
...
        "/teams/{teamId}/flagsets": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "flagsets"
                ],
                "summary": "List flagsets of a team",
                "parameters": [
                    {
                        "type": "string",
                        "description": "team id",
                        "name": "teamId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/model.APIResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/model.Flagset"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            },
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "flagsets"
                ],
                "summary": "Create a flagset under a team",
                "parameters": [
                    {
                        "type": "string",
                        "description": "team id",
                        "name": "teamId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "flagset",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.CreateFlagsetRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/model.APIResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/model.Flagset"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/model.APIResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/model.APIResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "model.APIResponse": {
            "type": "object",
            "properties": {
                "data": {},
                "errors": {
                    "type": "object",
                    "additionalProperties": {
                        "type": "array",
                        "items": {
                            "type": "string"
                        }
                    }
                },
                "message": {
                    "type": "string"
                },
                "success": {
                    "type": "boolean"
                }
            }
        },
...
        "model.CreateAPIKeyResponse": {
            "type": "object",
            "properties": {
                "apiKey": {
                    "type": "string"
                },
                "keyHash": {
                    "type": "string"
                }
            }
        },
...
        "model.CreateFlagsetRequest": {
            "type": "object",
            "properties": {
                "description": {
                    "type": "string"
                },
                "exporters": {
                    "type": "object"
                },
                "fileFormat": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "notifiers": {
                    "type": "object"
                },
                "pollingIntervalMs": {
                    "type": "integer"
                },
                "retrievers": {
                    "type": "object"
                }
            }
        },
...
        "model.Flagset": {
            "type": "object",
            "properties": {
                "apiKeyCount": {
                    "type": "integer"
                },
                "createdAt": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "exporters": {
                    "type": "object"
                },
                "fileFormat": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "notifiers": {
                    "type": "object"
                },
                "pollingIntervalMs": {
                    "type": "integer"
                },
                "retrievers": {
                    "type": "object"
                },
                "teamId": {
                    "type": "string"
                },
                "updatedAt": {
                    "type": "string"
                }
            }
        },
...
        "model.UpdateFlagsetRequest": {
            "type": "object",
            "properties": {
                "description": {
                    "type": "string"
                },
                "exporters": {
                    "type": "object"
                },
                "fileFormat": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "notifiers": {
                    "type": "object"
                },
                "pollingIntervalMs": {
                    "type": "integer"
                },
                "retrievers": {
                    "type": "object"
                }
            }
        },
...
    }
}